Don't put timestamp on generated file.
The timestamp changes leading to stupid problems with git checking out
carray related mako files after the generated files which will force a
recompilation and consequently a modified version of the carray files.
This is a nuisance and users may accidentally commit these irrelevant
changes.  I also explicitly set the version to 1.0a1.dev0 instead of it
being automatically normalized.